+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [1.5.0] - 2018-04-16
 
 ### Added
 
@@ -494,7 +494,8 @@ Initial release
 
 
 
-[Unreleased]: https://github.com/sdispater/pendulum/compare/1.4.4...1.x
+[Unreleased]: https://github.com/sdispater/pendulum/compare/1.5.0...1.x
+[1.5.0]: https://github.com/sdispater/pendulum/releases/tag/1.5.0
 [1.4.4]: https://github.com/sdispater/pendulum/releases/tag/1.4.4
 [1.4.3]: https://github.com/sdispater/pendulum/releases/tag/1.4.3
 [1.4.2]: https://github.com/sdispater/pendulum/releases/tag/1.4.2